Add I3MuxingReaderService, an extension of I3ReaderService for multi-stage processing.
I3MuxingReaderService takes the FileNameList form of I3ReaderService one step
further: instead of a single list of files, it takes multiple lists. The last of
the lists is the ``canonical'' stream; every time a physics frame is popped from
the canonical stream, the service searches through the other streams for frames
with the same DrivingTime and adds any new keys to the canonical frame. This can
be used e.g. to recover frame objects that were deleted in an earlier stage of
processing.

Since this recovery procedure is automatic, it may obviate the need to copy
lower-stage data into new files at every new processing stage. Instead,
higher-level files could become diffs against the previous stage, containing only
objects that were added or modified since the last stage. To that end,
I3MuxingReaderService keeps a copy of the frame injected into the module chain;
any frame can be turned into a diff by querying the service. This is implemented
in the I3FrameDiff module in phys-services; after I3FrameDiff, frames that passed
through the module chain unmodified contain only DrivingTime, which is needed to
unambiguously identify events.

The search is currently somewhat brain-dead; the service keeps a collection of
vanilla I3ReaderServices, each attached to one of the file lists, and repeatedly
calls PopMeta()/PopEvent() on each until the desired DrivingTime comes to the
front of the stream. This makes it necessary to still read the entire I3 file,
even if no frames are passed into the module chain. This would all be much more
efficient if there were a way to index into an arbitrary frame.

svn path:/highlander/trunk/; revision:168929