Change the way we do jni/java detection on Mac OS X
The "Apple Way" is to pass "-framework JavaVM" to the linker. Otherwise,
you may end up w/ an incomplete link line, which has been happening to us.
Now we just look for the jni includes, and use the frame work. This gives 
us a proper link.

This commit fixes the link issues on OS X. Tested on 10.5.8, 10.6.4 and
10.6.6.

This file needs to be re-organized, and the Apple bits need to use
CMAKE_FIND_FRAMEWORKS instead of the hard coded framework paths.

svn path:/highlander/trunk/; revision:169497