Monkeypatch a seg fault when using Icetray (in particular icetray-inspect) with Python 2.7.4 and newer.
Description of bug:
If the global logger was set from python, the global shared pointer in
libicetray.so would be destroyed during shared library unload in
__cxa_finalize(). At this point, however, the python interpreter has gone
away already, and the associated Py_DECREF() call will result in an
explosion. Instead, register an atexit() that turns logging off (actually
reverts to the default logger, which will be reinstantiated if more logging
occurs).

svn path:/highlander/trunk/; revision:168525