Remove the (unused) naked-streambuf constructor, fixing the build with Boost 1.55.
Background: boost::interprocess::buffer stream and friends now inherit privately from their underlying streambufs, and since access is not considered when resolving overloads, this makes the portable_binary_iarchive constructor ambiguous. If for some reason you need to bind an archive to a free-standing streambuf, wrap it in a dummy iostream.

svn path:/highlander/trunk/; revision:168561