setup "system include directories"
The reason why we have so many warning supression flags is that
large utilities like boost or python or root may have warnings that
are triggered by `-Wall`. These warnings then cause a flood of false
positives that distract from real warnings in our code. This commit
removes most of the warning supression flags, and declares the
include directories of `root`, `boost`, and `python` to be "system
include" directories.

If other utilities prove to be a problem with floods of warnings,
they can be added to the list, or the conditional removed out right,
marking all third party libs/includes as "system includes".

svn path:/highlander/trunk/; revision:169910