clean out cruft from the macro `i3_test-script()`
this cruft supported the old test framework and snowblower, and involved a
lot of writing and scraping of files. now, with ctest, this is just unused
code and bloats the macro making it harder to read. the macro 
`i3_test_executable()` contains similar cruft, but is a more complicated
macro and needs to be pruned more carefully.

svn path:/highlander/trunk/; revision:170294