Recursively cache parents of parents when merging frames
This fixes a severe performance regression introduced in r149410, which
attempted to fix ticket:1719. As a side-effect, though, it only cached direct
parents. This meant that the Q parent of a P frame would be cached, but GCD
frames would be read from the input file for every P frame. For a gzipped
Upgrade mDOM GCD, this caused steamshovel to take 30 seconds to move between
frames.

svn path:/highlander/trunk/; revision:170521