Fix ticket:2136.
When adding tray segments, I3Tray automatically propagates If into the
modules added by the segment. However, it used a single global variable
that could be overridden if that segment added another segment with a
different condition (or that had a keyword argument named "If"), resulting
in generalized space madness. Fix this by maintain a depth-indexed record
of conditions.

svn path:/meta-projects/combo/trunk/; revision:172193