Add `__bool__` operator to pybindings for python3 compatibility.
Effected classes are: `I3Double`, `I3UInt32`, `I3MCTree`, `I3FilterResult`, 
`I3Bool`, `I3Int`, and `ValidationReport`. This is nescessary because python2
uses `__nonzero__` for the binary operator, but python3 changed to `__bool__`.
All of these classes now have both operators defined, for compatibility with both 
languages.

Unit tests were added for all effeced classes.

In addition, it was noticed that the logic for `I3MCTree` was incorrect. It was changed 
so that `I3MCTree` will be false on an empty tree and true on a non-empty tree. 
It had mistakenly been set to the opposite. 

Also `test_tree_traversal` was enabled for `I3MCTree` as it had been mistakenly disabled 
due to a formatting error.

fixes #2341

svn path:/meta-projects/combo/trunk/; revision:175121