Removes `-ltinfo` flag from dataio-shovel to make it working again.
dataio-shovel is linked twice against different versions of libform and libncurses:
libform.so.5 and libncurses.so.5 from the system and libform.so.6 and libncurses.so.6 from the cvmfs.
libtinfo.so.5 from the system is always included, regardless of setting `-ltinfo` or not
However, explicitly including it results in a segfault of dataio-shovel whereas it runs fine in the other case (w/o the flag).
No idea why... see also #2408

svn path:/meta-projects/combo/branches/snobo/; revision:179003