Bug/svn calls in cmake, svn references in docs (#2480)
* remove svn calls from config.cmake and fix vars

Removes external calls to svn and replaces SVN_URL and SVN_REVISION with
git equivalents.

* fix up svn info in I3TrayInfo

I'm not nutty about this.

Puts GIT_URL in I3TrayInfo.svn_url
Puts GIT_REVISION in new I3TrayInfo.git_revision
Puts a magic number in I3TrayInfo.svn_revision

* fixes svn variables in templates

Replaces the variables SVN_URL and SVN_REVISION with their git
counterparts.

* do the same for docs

* do the same for parasitic builds

* do not add svn info to tarballs

* fixes the generation of the META_PROJET variable

* removes branch-metaproject

`branch-metaproject` was a utility to branch meta-projects. Useful in
the old svn repo w/ the old multi-project layout. Not useful with a flat
repo in git.

* removes review

`review` was a utility to give you a url to the trac page of the latest
commit of the current project. Useful for reviewing code on trac. Not
needed with github.

* sanitize the META_PROJECT variable

* updates the make_tarball.sh template

Adds the git revision and the correct compiler version to the tarball
ABOUT file.

* Compatibility with C++17 (#2474)

These changes will allow icetray to be compiled as C++17, they will not effect
its ability to compile with C++11 and it will not change that it is compiled as
C++11 by default.

* exclude .git directory from installation targets (#2473)

* remove G4DataQuestionaire from G4IceTopPhysicsList (#2476)

G4DataQuestionaire was removed from newer versions of Geant4 and it doesn't
appear to do anything.

* adding an initial CODEOWNERS file with a few names filled in. (#2477)

* updates parasitic metaproject doc for git

* removes svn mention in `make tarball` output

* removes the SVN Intro from the docs

* removes out of date "test drive"

* removes Commit Emails doc

Removes the .rst as it's no longer valid w/ the move to github. we can
always re-add it in the future if we use it for CI purposes

* updates tutorial doc with git

* add a warning about a section referencing SVN

The doc, Organization/Structure of Projects and Metaprojects, has a
secting describing the "traditional" SVN layout. This commit adds a
warning that the section is out-of-date with the move to GitHub, and
only applies to the historical svn archive.

* update svn references in docs

This commit updates some of the files in `docs/`. It removes references
to svn, and replaces them with git equivalents. In some places it fixes
nearby documentation.

* derives SHOVEL_VERSION from GIT_REVISION

* fixes more rst docs under cmake

* back out incomplete svn->git update in icetray

Co-authored-by: Kevin Meagher <11620178+kjmeagher@users.noreply.github.com>
Co-authored-by: Alex Olivas <aolivas@umd.edu>