Clean up steamshovel warnings (#2496)
* silences OpenGL deprecation warnings on macOS

Apple, has deprecated OpenGL on macOS, but gave us a define to globally
silence deprecation warnings.

* enables deprecation warnings in debug builds

* replaces auto_ptr with unique_ptr

* removes unused typedef

* fixes deprecation warnings (PyEval_InitThreads)

Extends an existing version check and moves PyEval_InitThreads() there.

PyEval_InitThreads() has been handled by Py_Initalize() since 3.7, was
deprecated in 3.9, and will be removed in 3.11.

* fixes deprecation warnings (QWheelEvent)

Replaces .x()/.y() with .position() (new in qt 5.14) via version check.

Outright replaces .delta() with .angleDelta(), as it was obsoleted in
qt 4.8

* fixes deprecation warnings (QString)

Updates a call to .split() with new semantics (new in qt 5.14) via
version check.

* fixes deprecation warnings (QColor)

Replaces .light() with .lighter(), deprecated in qt 4.8.

* fixes deprecation warnings (QTime)

Replaces a QTime based timer with a QElapsedTimer, as the former was
deprecated in qt 4.8.

* fixes deprecation warnings (qSort and QVector)

Replaces a QVector with a std::vector. The only reason a QVector was
being used here was because of the call to qSort() which qt has long
deprecated in favor of std::sort.