Set -fno-delete-null-pointer-checks when building with Clang.
Recent versions of clang (at least clang 11) optimize away null
pointer checks if they occur in the same context as code that
assumes pointers are non-null. This is particularly problematic
in some of the Boost serialization code, for instance at line
465 of archive/detail/oserializer.hpp, which dereferences (but
does not use) a potentially-null pointer for template resolution
purposes. Clang sees this -- which implicitly assumes the input
pointer is non-null -- and so optimizes away the following NULL
pointer check, resulting in serialization failures. Setting this
flag fixes 17 tests.

GCC can also do the same optimization but does not, for whatever
reason, in the problematic areas. We may need this flag for GCC
in future as well.