Updates to docs that are under cmake (#2613)
* Let the reader know doc is woefully out of date.

* Fixes duplicate targets

Unfortunately this forced to extend the entirety of the right-hand
column by 1.

* Fixes an external link to fftw

* Updates doc for Python3

* Overhauls the I3Test documentation

A rather large commit for documentation that isn't new, but it makes
more sense to do it this way.

 * fixes typos and spelling
 * indent the macros under "Build Configuration"
 * fix so other paragraph markup issues
 * use the `.. c:alias::` directive vice `.. c:macro::` in the macro
   definitions. The latter creates a new definition, and since these
   macros are defined in `i3test.h` we have an instant collision. There
   is no alias role in the sphinx cpp domain.
 * move some generically marked up "code" to more specific sphinx roles

* Improves the docs for our custom CMake macros

Still needs a lot of improvement, but at least now it renders some html
vaguely looking like source documentation.

I think we used the old sphinx CMake domain, but that got out of date
years ago at this point. Replaced the `.. cmake::` directives with the
more generic `.. object::` directive. We should look in to a modern
replacement for the CMake domain/extension. It seems we have a couple of
options. Not sure it's worth it though for this handful of macros.