CMake policy related housekeeping (#2598)
* Promotes policy settings to the top level.

Moves policy settings out of projects and into the top level
(`cmake/toplevel.cmake`). Sets them all to "new". CMP0046 will cause
(fixable) issues.

* Removes dependency on nonexistent target.

* Correct some file-based dependencies.

Using `add_dependencies()` to create a file-based dependency just
creates a dependency on a nonexistent target (which gets ignored).

This is now an error as of 505cd0e9.

* Removes some nonexistent dependencies.

`icepick`, `linefit`, and `gulliver-modules` don't have pybindings.

This was silently ignored prior to 505cd0e9.

* Adds a fake pybindings target to serialization.

With the policy switch in 505cd0e9 maybe a third of projects now error
out because they include `serialization` in their `i3_add_pybindings()`.
This is just a fake no-op target to silence those "nonexistent target"
errors. This is the lazy way of fixing that problem.

* Adds a fake pybindings target to CommonVariables.

Similar to 642d11ec, this adds a fake target to silence "nonexistant
target" errors cause by the flip in 505cd0e9. If the primary pybindings
had been named "CommonVariables" instead of (common_variables), this
wouldn't be necessary. But this patch is easier than renaming the
pybindings and fixing everything that relies on them.