delete very old code that linked to the wrong library
This code looked for the Capital letters `ARCH` in the build name and if found
it would add libgslblas to the linker. There are two things wrong with this,
first arch linux is lowercase and second libgslblas isn't a very good blas
implementation. I don't know when/why arch switched to being lower case so it
is hard to know how out of date this is. As stated in the documentation
Arch Linux users should just install the aur package `openblas-lapack`