[dataclasses] Remove base from ModuleKey
ModuleKey inherited from I3FrameObject, presumably to enable storage of
freestanding ModuleKeys in the frame. The pybindings did not expose this
inheritance, and no c++ code in icetray appears to place ModuleKeys
directly in the frame, so it is unlikely that this feature was ever
used. Removing the inheritance works around #2648, and also halves the
size of ModuleKey in memory.

Closes #2648.