Use leading underscore in python extension modules for normal importing
The standard convention for dealing with python extension modules that have the
same name as the regular module is to prefix the extension with an underscore.
IceTray seems to predate this convention and uses the same name for both modules.
This renders the extension modules inaccessable to python's normal importing
functionality. Instead it has a function called `load_pybindings()` which uses
imp.load_dynamic() to load the extension module by full path name. `imp` is
marked as depricated so this has to go.

This commit changes it so tha by default all pybindings will be build with
a leading underscore in the name. The macro  `I3_PYTHON_MODULE()` which
previously did nothing but call `BOOST_PYTHON_MODULE()` now adds and underscore
to the name of the module. Pybindings which don't have a competing python
module can still be build by using `BOOST_PYTHON_MODULE()` and adding the
option `NOUNDERSCORE` to `i3_add_pybindings()` in CMakeLists.txt.

There didn't seem to be any convention which projects used `I3_PYTHON_MODULE()`
and which ones used `BOOST_PYTHON_MODULE()`. This commit changes it so that all
modules requireing an underscore use I3 and all others use BOOST.

In principle this commit allows `load_pybindings()` to be dropped completely
and all python modules simply import the pybindings with `from _myproject
import *`. But that would increace ths size of the diff and take a lot of work.