disables broken test on specific github runner
This way the test will still fail locally. This shouldn't be taken as a
"let's just ignore this forever and never ever fix it" issue, or "the
runners must always be green!!!!11uno". The point is, we have a broken
test on a specific distro/OS. We disable it there, until it can be
properly fixed, and so it isn't holding up other commits. The fix to the
test will obviously remove the 'runner check', re-enabling the test.