Do less chisquare tests on the poission distribution for MT19937
As the docs for `scipy.stats.chisquare()` it is only supposed to reliably work
with with frequencies well above 5. The poisson distribution test was giving
frequencies well below that. This just drops the tests with poisson parameters
so it is only done from .1 to 8 rather than all the way to 10.

fixes #2271