Make tableio and hdfwriter tests pass if simclasses and recclasses absent.
In a few places, we have imports of simclasses and recclasses that are not
strictly necessarily. hdfwriter-book imports recclasses for no particular
reason, and one tableio test uses simclasses sometimes. Fix this by
removing the unneeded import and by making the simclasses part of the
tableio test conditional on the presence and importability of simclasses.