python tests - eliminate races with output files
The "tests" in this commit all used "output.i3.gz" as an output file
name causing obvious races when tests are run in parallel. Do away with
the named output file, and use a NamedTemporaryFile() context manager
instead, so the temp file is auto deleted when it falls out of scope.

These changes could be abstracted further and put in to
I3Test::TestExampleScripts, along with defining input files. Also
updated related documentation.

See #2911