Prevent users from accidentally putting None in I3Frame from Python (#2994)
* Prohibit I3Frame.Put(str, NoneType)

* Delete pulses in tensor-of-interia test

Putting `None` in the frame is no longer supported.  My best guess
is that the test orginally tried to delete OfflinePulses from the
frame, but since this was mixed in the from the Q frame, the deletion
didn't stick.  Overwriting it with `None` in the P frame would have
prevented I3Module from mixing it in again, but is otherwise a
confusing thing to do.

* Avoid trying to formal nullptr in I3FrameTest