Set ATWD flag based on pulse width in SuperDST (#3017)
I3RecoPulse has separate flags for LC, ATWD, and FADC, while SuperDST
only stores the LC bit.  When decoding, it set pulse flags to
LC|ATWD|FADC if the LC bit is set, and FADC otherwise.  This is
still the default behavior, but if I3SuperDST.use_width_for_atwd_flag
is set to True, trailing LC pulses will be marked FADC-only if their
width is not <= to the width of the first pulse (always from the
ATWD in current WaveDeform).

Fixes #2721.