add toggle for `dataio-shovel`'s background color (#3122)
* add toggle for dataio-shovel's background color

Add a function and matching key binding (t) to toggle the background
color of `dataio-shovel` between its default black and the terminal's
default. This means that in a black-on-white terminal, `dataio-shovel`
can have a white background. Think of it as a black/transparent toggle,
vs a black/white toggle.

Note that your terminal's theme will determine how good this actually
looks.

Closes #3119

* turn off the tilde EOF marks in less

* clean up help message. keybinds aren't configurable