fix boost detection with parasitic builds (#3230)
* fix boost detection for parasitic builds

We kind of abuse cmake and our build system with parasitic builds by
slupring in the CMakeCache.txt of the "host" metaproject. Unfortunately
we store the "imported targets" provided by Boost in BOOST_LIBRARIES.
This confuses CMake in parasitic builds because `find_package(Boost)`
was never run, and it thinks that it needs to link a library with a
filename of Boost::system (for example).

This commit special-cases Boost and "redetects" it instead of slurping
from the "host" CMakeCache.txt.

Closes #3228

* update release notes