clean up new noqas in steamshovel (#3292)
* steamshovel - fix noqa: PLW3301 with a comment

generally you want to flatten calls to max() or min() but you can't mix
a floats and tuples. therefore it's ok to have a nested max() call here,
as the inner max() can operate on the rgb-3-tuple, and the outer one can
then operate on that result and the normalized brightness.

* steamshovel - fix "noqa: B012"

a return() in a `finally:` will suppress all exceptions, and that's not
what we want. move the return out.