Remove naive datetime() objects (#3304)
* Remove naive datetime() objects

Python is transitioning away from naive datetime objects, objects that
don't have a timezone. Naive datetimes are danerous because they are
usually used for localtime but don't behave correctly for localtime.
For the most part we seem to use naive datetime for utc which is less
bad. This modifies the pybindings for I3Time so that I3Time.date_time
returns an object with the tzinfo set to UTC and `make_I3Time()` only
accepts objects with tzinfo set to UTC. All instances of datetime()
in examples, scripts, and tests have been switched to set UTC, and
"DTZ" has been enabled for ruff.

* remove unneeded debug statement