Fix #3172 (#3309)
* clear error state after bp::extract

`I3Module::GetParameter()` tries twice to fetch a parameter, first by
type, and if that fails then by name. We've never cleared the error
state before the second fetch attempt. According to the Python docs
"additional calls into the Python/C API may not behave as intended and
may fail in mysterious ways."

We don't need to bother clearing the error state if the second attempt
fails, as we're just going to re-throw and crash anyway.

* fix default value of IterationTweakService in I3IterativePandelFitter

The defaualt value is this case should be a Python `None`, as this is
what I3Configuration tests on for a "thing-not-found".