Remove Uber Header (I3.h) (#3151)
* Remove Uber Header (I3.h)

For a long time we had a piece of technical debt called the uber header. This was a header that was included in every file in the repo by using `-include` in COMPILE_FLAGS. It was also created using cmake's `configure_file()` even though it didn't have any macro replacements. The file didn't do much, it mostly was just a convoluted way to include `pybindings.hpp` in python bindings and a few global defines. A lot of the global defines are work arounds that are decades old.

This patch removes the uber header. For all files which need `pybindings.hpp`, it is now included in the normal way. The defines that are still relevent were moved into cmake logic using `add_compile_definitions()`. And if we ever need any more configured defines they can be put into `workspace_config.h` instead.

* fix conflict
* remove I3_UBER_HEADER that slipped into merge
* fix bad merge-conflict-resolution

---------

Co-authored-by: nega <nega0@users.noreply.github.com>