Add SPDX tags (#3387)
* Add SPDX tags to IceTray

The goal of this is to get our software complient with reuse https://reuse.software/.
Copyright notices are kept intact and only modified to be complient with reuse formatting.
Copyright notices are added where none exist as `2024 The IceTray Contributors`.
Files with licences have the correct `SPDX-License-Identifier` identifer added.
Files without licenses are licensed with `BSD-2-Clause`.

Two custom licences were added: `LicenseRef-Esquivel` which is just a short very permisive
licence that was clearly just written by the author of the file,
and `LicenseRef-Geant4` because apperently we copied I3Units from Geant4.

The following files had licence declerations in the file and the corresponding SPDX
were added

BSD-3-Clause:
* icetray/private/icetray/crc32.c

LicenseRef-Geant4:
* icetray/public/icetray/I3Units.h

BSL-1.0:
* icetray/public/icetray/counter64.hpp
* icetray/public/icetray/i3_extended_type_info.h
* icetray/public/icetray/Singleton.h
* icetray/public/icetray/python/std_vector_indexing_suite.hpp
* icetray/public/icetray/python/std_map_indexing_suite.hpp
* icetray/public/icetray/python/list_indexing_suite.hpp

LicenseRef-Esquivel:
* icetray/public/icetray/python/function.hpp

* fix tag for I3Units.h

* fix copyright statement in Version.h

* license I3Units as BSD-2-Clause

Based on comments from nwhitehorn remove comments pretaining to Geant4
and copyright as icetray contributiors and license as BSD-2-Clause.
According to nwhitehorm 'I don't think (a) it is meaningfully copyrightable
since it is a list of facts and (b) that it has much in common with the
original header anymore. I would prefer just to strip the GEANT4 comments
from that file.'