Add SPDX tags (#3435)
The goal of this is to get our software complient with reuse https://reuse.software/.
Copyright notices are kept intact and only modified to be complient with reuse formatting.
Copyright notices are added where none exist as 2024 The IceTray Contributors.
Files with licences have the correct SPDX-License-Identifier identifer added.
Files without licenses are licensed with BSD-2-Clause.

The only file not written by an IceCube member is Miniball.h which is GPL-2.0-or-later