Add SPDX tags (#3439)
The goal of this is to get our software complient with reuse https://reuse.software/.
Copyright notices are kept intact and only modified to be complient with reuse formatting.
Copyright notices are added where none exist as 2024 The IceTray Contributors.
Files with licences have the correct SPDX-License-Identifier identifer added.
Files without licenses are licensed with BSD-2-Clause.

Three files were written by non-icecube members which are all licensed with BSL-1.0

- utility/pretty.cmake which
- tool-patches/common/I3/intel_compatibility.h
- tool-patches/boost-new/boost/shared_ptr.hpp

Co-authored-by: nega <nega0@users.noreply.github.com>