Add SPDX tags (#3440)
* [docs] Add SPDX tags

The goal of this is to get our software complient with reuse https://reuse.software/.
Copyright notices are kept intact and only modified to be complient with reuse formatting.
Copyright notices are added where none exist as 2024 The IceTray Contributors.
Files with licences have the correct SPDX-License-Identifier identifer added.
Files without licenses are licensed with BSD-2-Clause.

The only file not written by icecube was basedocs/icetray/todo.py which was written by
sphinx which is BSD-2-Clause

the tag REUSE-IgnoreStart is used in several files so that copyright informaiton
in the documentation isn't confused by reuse

* align indentation, fix typo

* fix unintentional changes

* fix artifact from merge

---------

Co-authored-by: nega <nega0@users.noreply.github.com>