SPDX tags for infrastructure files. (#3454)
* SPDX tags for infrastructure files.

With the commit icetray-public should be reuse complient, excepting that it
has a differnet README file

dep5 file was used to specify licenses for non-text files. I tried not to go
too overboard with the globs in that file.

tree.pdf was deleted, the code we use is GPL but the license of the pdf file
is unclear. In addition we really don't need to carry around a pdf file in our
repo just to documnet a header library which is easy to google

* fix some issues raised by nega