Run mypy in CI (#3480)
This PR adds mypy static type checking as a CI step, using the same environment used to generate the type stubs. This makes it slightly simpler to deal with libraries that are not built on all platforms.

Checking the Python components of IceTray libraries is relatively straightforward via e.g. `mypy -p icecube.icetray`. Scripts are slightly trickier, as mypy _really_ wants everything to be a module, i.e. have a unique name. To work around this, the `mypy-scripts` cmake target creates a directory _mypy_scripts with symlinks to the `PYTHON_DIR` of each project, with hyphens replaced with underscores. `mypy` can be convinced that this is a namespace package and check it in one go. This makes it nontrivial to find the file that led to an error (e.g. when mypy complaints about _mypy_scripts/simprod_scripts/resources/examples/proposal_alternate_cross_sections.py, it means simprod-scripts/resources/examples/proposal_alternate_cross_sections.py), but is the least bad approach I could come up with.

The checks currently fail, pending #3414, #3415, #3416, #3417, #3418, and #3419. Once those are merged, the only remaining failures should be those explicitly ignored in pyproject.toml.

* Check Python modules with mypy
* ci: pin stubgen image
* Specify mypy target
* run mypy serially
* Ignore mypy cache
* run mypy in separate job
* pybindings: ensure dataio is imported
* Add mypy ignores to pyproject.toml

  formerly file-level mypy: ignore-errors comments

* Run mypy check against merge commit for PRs
* ci: bump stubgen image
* stubs: update README
* rpdf: set defaults for optional kwargs
* stubs: generate for cb37c4f049f022efa6753d01e776a9683395c196

  Co-authored-by: Jakob van Santen <jvansanten@gmail.com>

* finallevel_filter_diffusenumu: lib shiny for mypy

  ConfigParser.readfp was deprecated in Python 3.2 and finally removed
  in 3.12

* simprod-scripts: lib shiny for mypy
* icetop_Level3_scripts: shiny for mypy
* ml_suite: lib shiny for mypy
* ml_suite: scripts shiny for mypy

  Fail loudly if I3_BUILD is not defined

* ml_suite: ignore script with undefined name error
* HiveSplitter: ignore broken scripts
* Disable on check on push
* pybdt: link python bits at configure time

  rather than deferring to a target.  This ensures that lib is populated
  after cmake is run, just like for other projects.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>