mark icetray.logging.log_fatal as NoReturn
This tells type checkers that branches including a log_fatal are
never followed, so e.g. `if foo is None: log_fatal("")` narrows
the type of `foo` to exclude NoneType.

Along the way, fix a case in radcube.SimulateFromStar where Configure()
log_fatal()ed unconditionally as far as mypy was concerned.