Fix copyright and license statements in simulation projects (#3546)
* Fix copyright and license statements in simulation projects

Unlike previous PRs the only fixes existing copyright statements so they
are compatable with reuse. SPDX license tags are added to files which
already contain license text. As we have not premission to relicense
from all authors on these projects there are no licenses aded to files
which don't have any license information.

This covers all simulation projects not in #3542 #3543 #3545

* resort

---------

Co-authored-by: nega <nega0@users.noreply.github.com>