Add Integer32 to I3RandomService (#3642)
* Add Integer32 to I3RandomService

This has been sometihg that I have been stepping around for a while.
Our interface takes a 32-bit max as an exclusive range, which means it
is impossible to ask for an range that spans the entire 32-bit integer.
This isn't a big deal most of the times since most of our RNG useage is
simulating particles with floats. But we use our RNG to seed CLSim's MWC
RNG which needs a 32-bit integer. With this patch I can reproduce
CLSims results with out having to do anything silly to compenstate.

* stubs: generate for 13ceb0cc71490c58087589f3e67902a813dedca7

Co-authored-by: Kevin Meagher <11620178+kjmeagher@users.noreply.github.com>

* fix incorrect name in previous commit

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: nega <nega0@users.noreply.github.com>