Switch tableio tests from pytables to h5py (#3726)
* Switch tableio tests from pytables to h5py

pytables is fine when it is compiled correctly but consistently doing
that consistently for CI is proving challenging. pytables is very slow
at providing precompiled wheels for new versions of python and even then
they tend to have incompatablilty issues with installed libhdf resulting
in segfaults.

This patch switches all of our tests to use h5py which is more reliable
for CI. There are still pleanty of usecases for tables in our code base
but they aren't used for unit tests.

* update python imports in tests

- reorder imports, placing hdfwriter after h5py
- remove test skip conditions
- fail tests on failure to import

These changes reduce segfaults and HDF5 version conflicts.

---------

Co-authored-by: nega <nega0@users.noreply.github.com>