Make tests clang-tidy clean (#3549)
* clean up tabs and related mess

* remove unused includes

* fix tests broken by prior commit

* improve const-correctnes in a few macros

* dataio - fix tabs in test file

* dataio - simplify test

One can construct an I3Double() from an int, so there's no reason to
have a double as the loop counter. Also, there's no reason to construct
an I3DoublePtr() twice.

* dataio - clean up tabs and indentation

* dataio - clean up includes

* dataio - clean up loops

The loop counter is immediately cast to a string, so there's no reason
to use a double as the counter. Also, change said cast to std::to_string

* dataio - use nullptr and initialize a stat struct

* serialization - nolint a warning about non-optimal padding

This is a dummy struct used in several of the tests that is not
optimally laid out. We want to maintain as much parity with the original
boost sources to aid in any future back-porting. Saving a few bytes in a
few short-run-time tests isn't worth it.

* serialization - nolint a dead store

I have no idea why clang thinks this is a dead store, but here we are.

* lilliput - clean up headers

* lilliput - make loop counter an int

* gulliver - clean up gradient test

* DeepCore_Filter - make header file a header file

* DeepCore_Filter -- cleanup test helpers

* DOMLauncher - change loop counter from double to uint

* radcube - remove dead store

* earthmodel-service - add test for function result

this eliminates a dead store

* dataclasses - silence warning triggered from deep inside boost

* g4-tankresponse - cleanup the triangle test

* dataio - undo change of lexical_cast to to_string