Reduce rounding error in ExtrudedPolygon::GetIntersection (#3776)
As pointed out in #3774, ExtrudedPolygon::GetIntersection() suffers
from significant roundoff error when the x component of the direction
is nearly, but not quite entirely, 0.  Switch to a more expensive,
but more numerically stable, calculation that interpolates smoothly
between the x and y projections.

Also, add a test to verify that this handles the formerly pathological
cases well.

Fixes #3774