Documentation build process cleanup (#3915)
* fix typos

* add undocumented breathe debug flags to sphinx config

* add I3Logging header to files missing it

Among doxygen's many... foibles... is that it can get confused by 1)
function-like macros, and 2) transitive includes. This patch adds
`SET_LOGGER` to the doxygen's "symbol exclude list" and `#include
"icetray/I3Logging.h"` to each file that calls `SET_LOGGER`. Both steps
are necesasry for doxygen to ignore `SET_LOGGER`. The reduction in
warnings that this provides results in about a 40% decrease (by line) in
the build process output.

* add std and std::* to doxygen symbol excludes

* doxygen doesn't like function-like variable initialization

* Swap the position of of attribute:unused to doxygen's liking

* clean up clsim doc warnings

This patch adds excludes for some clsim sub-directories, and OpenCL
headers. This reduces warnings in the doc build process by about 6100
lines.

* remove unused extensions

We haven't used them for a couple of years, and they were old and crusty
when they were disabled. Also removes our copy of the sphinx "todo"
extension that we don't use and have been carrying around.

Closes #3713

* lint the docs-build script and sphinx config

Crank up ruff linting on `build-docs` and the sphinx config. Basically,
we add "all" warnings, and ignore some that would be awkward to fix
(ie. subprocess input validation), or unnecessary in the context of a
standalone script.

* add a codespell exclusion

* tweak log messages

* tweak help message

* bump the copyright date in the HTML footer

* cmake - update ruse_check.py's ignore list