Remove All Star Imports (#3902)
* Remove All Star Imports

* only import clsim stuff and gsl rangom if they were compiled

* remove bitsets from frame_object_diff

I'm not sure what they are even supposed to do

* stubs: generate for 7a3509cf875b8ccc4f902c7451cc7c8dd3c7ff01

Co-authored-by: Kevin Meagher <11620178+kjmeagher@users.noreply.github.com>

* PyBindings for Conditionally compiled items are conditionally imported

This is done by exporting a module atribute with the same name as the
compiler definition. The module is attribute is checked before importing
the pybinding.

Also fix some other issues caught by mypy

* stubs: generate for fb35e5200e0a09c79dde1fc6eae8ee2c866af8a8

Co-authored-by: Kevin Meagher <11620178+kjmeagher@users.noreply.github.com>

* remove pybindings for SiPMResponseSimulator and level3-filter-muon

Neither of these project had any pybindings but they were using the
pybindings to load the shared library. This deletes the pybindings
and replaces the import of the pybindings in __init__ with an
icetray.load() call.

Add newlines to the end of file which were eronously dropped in the
initial commit

* stubs: generate for 9096ca302cd7ad529686360ecfbf8a8cf900e8b7

Co-authored-by: Kevin Meagher <11620178+kjmeagher@users.noreply.github.com>

* MuonGun - update ExtrudedPolygon import

* simclasses - remove EPOS import

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: nega <nega0@users.noreply.github.com>