Remove the last bits of rootcint support (#3935)
* remove CMake USE_CINT and C++ I3_USE_CINT

* cmake - remove boost::shared_ptr wrapper

This wrapper hid b::shared_ptr from rootcint by declaring it as an empty
struct.

* cmake - remove __CINT__ check from intel_compatibility.h

This was to merely set a warning suppression pragma for an old version
of the Intel compiler

* icetray - remove __CINT__ checks

* clean up header fallout

* dataclasses - remove __CINT__ checks

* mue -- remove __CINT__ checks

* phys-services -- remove __CINT__ checks

* ml_suite -- remove __CINT__ checks

* icetray - clean up __CINT__ checks in name_of()

Removes the __CINT__ checks and removes the wrapper that injected
`icetray::name_of()` into the `I3` namespace. Cleans up the relevant
call sites.

* deleted old root scripts

---------

Co-authored-by: kotoyo <6904302+kotoyo@users.noreply.github.com>