Remove triangle.c (#3998)
we used to use triangle.c to implement Delaunay Triangle grid. Over 7
years ago we switched to using the bowyer implementation.
triangle.c was left as a compile time option with the default off.
triagnle.c was written in a very old style and generates a lot of
warnings which are a burdent on the icetray team. Unless we really need
to match the really old data it is time to delete this.